requirements.txt
standard-aifc
opencv-python
opencv-contrib-python
pytesseract
beautifulsoup4==4.10.0
bidict==0.21.3
certifi==2021.10.8
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
comtypes==1.1.10
decorator==5.1.0
future==0.18.2
geocoder==1.38.1
idna==3.2
jsonlib-python3==1.6.1
loc==0.0.2
MouseInfo==0.1.3
pillow==11.0.0
psutil==6.1.0
PyAudio==0.2.14
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
pyjokes==0.6.0
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.28
pyttsx3==2.90
pytube==11.0.1
pytweening==1.0.4
pywin32==308
ratelim==0.1.6
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
urllib3==1.26.7
wikipedia==1.4.0
